+++ libs/common/src/lib/search-widget/search-configuration/search-box-form/search-box-form.component.ts
@@ -86,6 +86,8 @@ export class SearchBoxFormComponent implements OnInit, OnDestroy {
     prependTheQuery: new FormControl<boolean>(false, { nonNullable: true }),
     queryPrepend: new FormControl<string>('', { nonNullable: true, updateOn: 'blur' }),
     rephraseQuery: new FormControl<boolean>(false, { nonNullable: true }),
+    useRephrasePrompt:  new FormControl<boolean>(false, { nonNullable: true }),
+    rephrasePrompt: new FormControl<string>('', { nonNullable: true }),
     generateAnswerWith: new FormControl<'only-semantic' | 'semantic-and-full-text'>('semantic-and-full-text', {
       nonNullable: true,
     }),
@@ -117,6 +119,12 @@ export class SearchBoxFormComponent implements OnInit, OnDestroy {
   }
   get prependTheQuery() {
     return this.form.controls.prependTheQuery.value;
+  }
+  get rephraseQueryEnabled() {
+    return this.form.controls.rephraseQuery.value;
+  }
+  get useRephrasePromptEnabled() {
+    return this.form.controls.useRephrasePrompt.value;
   }
 
   ngOnInit() {